multi-modal head gate

+++ recognition/model/dynamicvit_legacy.py
@@ -81,7 +81,7 @@ class AVnet_Dynamic(nn.Module):
                     # TopK selection
                     # num_keep_node = int(self.num_patches * self.token_ratio[p_count])
                     # threshold selection
-                    num_keep_node = torch.argmax((values < self.threshold).float())[0].item()
+                    num_keep_node = torch.argmax((values < self.threshold).int())[0].item()
                     keep_policy = indices[:, :num_keep_node]
 
                     keep_token.append(num_keep_node / indices.shape[1])
